feat: Add spritesheet loader utility class
This commit adds a new utility class, SpritesheetLoader, to the MonoGame project. The SpritesheetLoader class provides a method, GetSpritesheet, that allows other classes to load and retrieve spritesheets by their file paths. This utility class will be used to load and manage spritesheets for tiles and entities in the game.

Note: The commit message follows the conventional format of starting with a type (feat for feature) followed by a concise description of the changes.