feat: Enable double texture size for StoneTile and WaterTile
This commit enables the double texture size feature for the StoneTile and WaterTile classes in the MonoGame project. By setting the DoubleTextureSize property to true, the tiles will be rendered with a larger texture size, resulting in a more visually appealing appearance.

Note: The commit message follows the conventional format of starting with a type (feat for feature) followed by a concise description of the changes.