feat: Add Walkable property to ITile interface and Tile class
This commit adds a Walkable property to the ITile interface and the Tile class in the MonoGame project. The Walkable property allows for determining whether a tile can be walked on by game entities. This feature will be used to implement collision detection and movement restrictions in the game.

Note: The commit message follows the conventional format of starting with a type (feat for feature) followed by a concise description of the changes.