refactor: Update MonoGame assembly version and source links
This commit updates the MonoGame assembly version and source links in the project. The assembly version is updated to "1.0.0+ff525dbe7a25a6681c8c0da649e0c3048426b056" in the MonoGame.AssemblyInfo.cs file. Additionally, the source link in the MonoGame.sourcelink.json file is updated to point to the commit "ff525dbe7a25a6681c8c0da649e0c3048426b056" on the GitHub repository. These changes ensure that the correct version and source code are referenced in the project.