feat: Update Camera class to use public property for ScaleFactor
This commit updates the Camera class to use a public property for the ScaleFactor instead of a private readonly field. This allows external code to access and modify the ScaleFactor as needed. The property is now named ScaleFactor and has a default value of 3f. This change improves the flexibility and usability of the Camera class.