feat: Add SlotUserInterfaceComponentController.ResetSlots method
This commit adds the ResetSlots method to the SlotUserInterfaceComponentController class. The method is called when a mouse button is released in the LevelEditor input event channel. It resets the SourceSlot and DestinationSlot properties to null, clearing the selected slots. This change improves the functionality and user experience of the SlotUserInterfaceComponentController.